updated controllers

+++ kolplace - Final React Project SoftUni/server/src/controllers/shoppingCartController.js
@@ -6,6 +6,9 @@ const { isAuthenticated } = require("../middlewares/isAuthenticated");
 
 shoppingCartController.get("/shopping-cart/:id", isAuthenticated, async (req, res) => {
     try {
+        if (req.decToken) {
+            return res.status(401).json({ expMessage: "Your session has expired, you have to login again!" });
+        }
         const cart = await shoppingCartService.getById(req.params.id);
         res.status(200).json(cart);
     } catch (error) {
@@ -15,6 +18,9 @@ shoppingCartController.get("/shopping-cart/:id", isAuthenticated, async (req, re
 
 shoppingCartController.post("/shopping-cart/:id", isAuthenticated, async (req, res) => {
     try {
+        if (req.decToken) {
+            return res.status(401).json({ expMessage: "Your session has expired, you have to login again!" });
+        }
         const updatedCart = await shoppingCartService.addToCart(req.params.id, req.body);
         res.status(200).json(updatedCart);
     } catch (error) {
@@ -24,6 +30,9 @@ shoppingCartController.post("/shopping-cart/:id", isAuthenticated, async (req, r
 
 shoppingCartController.put("/shopping-cart/:id", isAuthenticated, async (req, res) => {
     try {
+        if (req.decToken) {
+            return res.status(401).json({ expMessage: "Your session has expired, you have to login again!" });
+        }
         console.log(req.body);
         const updatedCart = await shoppingCartService.removeProduct(req.params.id, req.body);
         res.status(200).json(updatedCart);
@@ -34,6 +43,9 @@ shoppingCartController.put("/shopping-cart/:id", isAuthenticated, async (req, re
 
 shoppingCartController.delete("/shopping-cart/:id", isAuthenticated, async (req, res) => {
     try {
+        if (req.decToken) {
+            return res.status(401).json({ expMessage: "Your session has expired, you have to login again!" });
+        }
         const updatedCart = await shoppingCartService.emptyCart(req.params.id);
         res.status(200).json(updatedCart);
     } catch (error) {
